chore: restructure code - test of dashboard views
